[PC-538] Repository에서 안드로이드 의존성 ImageResizer로 분리 및 ProfileRepositoryTest fix
